Init data into database

+++ server/src/main/java/com/org/controller/BookingController.java
@@ -53,14 +53,14 @@ public class BookingController {
 
 	@GetMapping("/searchBooking/{id}")
 	@ExceptionHandler(RecordNotFoundException.class)
-	public ResponseEntity<?> searchBookingByID(@PathVariable("id") BigInteger bookingId) {
+	public ResponseEntity<?> searchBookingByID(@PathVariable("id") Long bookingId) {
 
 		return bookingService.findBookingById(bookingId);
 	}
 
 	@DeleteMapping("/deleteBooking/{id}")
 	@ExceptionHandler(RecordNotFoundException.class)
-	public void deleteBookingByID(@PathVariable("id") BigInteger bookingId) {
+	public void deleteBookingByID(@PathVariable("id") Long bookingId) {
 
 		bookingService.deleteBooking(bookingId);
 	}
