+++ server/src/main/java/com/org/dao/BookingDao.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.org.model.Booking;
 
 @Repository
-public interface BookingDao extends CrudRepository<Booking, BigInteger> {
+public interface BookingDao extends CrudRepository<Booking, Long> {
 
 }
