+++ server/src/main/java/com/org/dao/ScheduledFlightDao.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.org.model.ScheduledFlight;
 
 @Repository
-public interface ScheduledFlightDao extends CrudRepository<ScheduledFlight, BigInteger>{
+public interface ScheduledFlightDao extends CrudRepository<ScheduledFlight,Long>{
 
 }
