+++ server/src/main/java/com/org/service/BookingServiceImpl.java
@@ -61,7 +61,7 @@ public class BookingServiceImpl implements BookingService {
 	 */
 
 	@Override
-	public String deleteBooking(BigInteger bookingId) {
+	public String deleteBooking(Long bookingId) {
 
 		Optional<Booking> findBookingById = bookingDao.findById(bookingId);
 		if (findBookingById.isPresent()) {
@@ -89,7 +89,7 @@ public class BookingServiceImpl implements BookingService {
 	 * find booking by ID
 	 */
 	@Override
-	public ResponseEntity<?> findBookingById(BigInteger bookingId) {
+	public ResponseEntity<?> findBookingById(Long bookingId) {
 		Optional<Booking> findById = bookingDao.findById(bookingId);
 		try {
 			if (findById.isPresent()) {
