+++ server/src/main/java/com/org/service/ScheduledFlightService.java
@@ -12,11 +12,11 @@ public interface ScheduledFlightService {
 
 	public ScheduledFlight modifyScheduledFlight(ScheduledFlight scheduledFlight);
 
-	public String removeScheduledFlight(BigInteger id) throws RecordNotFoundException;
+	public String removeScheduledFlight(Long id) throws RecordNotFoundException;
 
 	public Iterable<ScheduledFlight> viewAllScheduledFlights();
 
-	public ScheduledFlight viewScheduledFlight(BigInteger id) throws ScheduledFlightNotFoundException;
+	public ScheduledFlight viewScheduledFlight(Long id) throws ScheduledFlightNotFoundException;
 	// boolean cancelBookings(BigInteger flightId) throws RecordNotFoundException;
 
 }
